+++ lib/web/fetch/request.js
@@ -11,7 +11,7 @@ const {
   isValidHTTPToken,
   sameOrigin,
   normalizeMethod,
-  makePolicyContainer,
+  EnvironmentSettingsObject,
   normalizeMethodRecord
 } = require('./util')
 const {
@@ -27,7 +27,6 @@ const {
 const { kEnumerableProperty } = util
 const { kHeaders, kSignal, kState, kGuard, kRealm, kDispatcher } = require('./symbols')
 const { webidl } = require('./webidl')
-const { getGlobalOrigin } = require('./global')
 const { URLSerializer } = require('./data-url')
 const { kHeadersList, kConstruct } = require('../../core/symbols')
 const assert = require('node:assert')
@@ -40,20 +39,6 @@ const requestFinalizer = new FinalizationRegistry(({ signal, abort }) => {
 })
 
 let patchMethodWarning = false
-
-class EnvironmentSettingsObjectBase {
-  baseUrl = getGlobalOrigin()
-
-  get origin () {
-    return this.baseUrl?.origin
-  }
-
-  policyContainer = makePolicyContainer()
-}
-
-class EnvironmentSettingsObject {
-  settingsObject = new EnvironmentSettingsObjectBase()
-}
 
 // https://fetch.spec.whatwg.org/#request-class
 class Request {
