+++ lib/web/fetch/util.js
@@ -1569,6 +1569,28 @@ function utf8DecodeBytes (buffer) {
   return output
 }
 
+class EnvironmentSettingsObjectBase {
+  #baseUrl
+
+  get baseUrl() {
+    return (this.#baseUrl ??= getGlobalOrigin())
+  }
+
+  set baseUrl(baseUrl) {
+    this.#baseUrl = baseUrl
+  }
+
+  get origin () {
+    return this.baseUrl?.origin
+  }
+
+  policyContainer = makePolicyContainer()
+}
+
+class EnvironmentSettingsObject {
+  settingsObject = new EnvironmentSettingsObjectBase()
+}
+
 module.exports = {
   isAborted,
   isCancelled,
@@ -1620,5 +1642,6 @@ module.exports = {
   createInflate,
   extractMimeType,
   getDecodeSplit,
-  utf8DecodeBytes
+  utf8DecodeBytes,
+  EnvironmentSettingsObject
 }
